+++ src/react/components/positioning/Positioning.js
@@ -240,21 +240,24 @@ class Positioning extends PureComponent {
 	 */
 	calculateMaxHeight() {
 		const popup = this;
-		const { viewportGutter, parentRef, triggerDropdownGap } = popup.props;
+		const { viewportGutter, parentRef, triggerDropdownGap, maxHeight } = popup.props;
 		const triggerDOM = parentRef.firstChild;
 
 		if (verge.inViewport(triggerDOM)) {
 			if (!popup.props.isNotResponsive && isNarrowViewport()) {
+				const viewportH = verge.viewportH();
 				popup.setState({
-					maxHeight: verge.viewportH(),
+					maxHeight: Math.min(viewportH, maxHeight),
 				});
 			} else {
 				const triggerRect = triggerDOM.getBoundingClientRect();
 				const spaceAboveTrigger = calcSpaceAbove(triggerRect);
 				const spaceBelowTrigger = calcSpaceBelow(triggerRect);
+				const availableSpace = Math.max(spaceAboveTrigger, spaceBelowTrigger) - viewportGutter - triggerDropdownGap;
+				const calculatedHeight = maxHeight ? Math.min(availableSpace, maxHeight) : availableSpace;
 
 				popup.setState({
-					maxHeight: Math.max(spaceAboveTrigger, spaceBelowTrigger) - viewportGutter - triggerDropdownGap,
+					maxHeight: calculatedHeight
 				});
 			}
 		}
@@ -326,6 +329,11 @@ Positioning.propTypes = {
 	onVisible: PropTypes.func,
 	/** Setting to true will for the dropdown to be as wide as the trigger. */
 	isTriggerWidthMatched: PropTypes.bool,
+	/**
+	 * Setting a number here will force the maximum height of the child to be the number provided (in pixels) if the viewport is too big.
+	 * When the viewport is smaller than this number, it still shrinks, but never grows beyond that number.
+	 */
+	maxHeight: PropTypes.number
 };
 
 Positioning.defaultProps = {